apps/domain-analyzer/requirements.txt
newspaper4k>=0.9.3
psycopg2-binary>=2.9.0
requests>=2.31.0
typing-extensions>=4.0.0